chore(frontend): move all other components to new composites folder
