+++ apps/frontend/src/components/composites/Hero.tsx
