+++ apps/frontend/src/components/composites/NavBarAuthSection.tsx
